moved utils into CalendarDatePicker folder and added dateIsSelectable function
